Fixed chart with comparator. Made more clear error handling with sharedFlow.

+++ app/src/main/java/com/dev/agalperin/domain/model/Point.kt
@@ -1,6 +1,6 @@
 package com.dev.agalperin.domain.model
 
 data class Point(
-    val x: Double,
-    val y: Double
+    val x: Float,
+    val y: Float
 )
